+++ src/Confluent.SchemaRegistry/Rest/IRestService.cs
@@ -29,15 +29,13 @@ namespace Confluent.SchemaRegistry
     internal interface IRestService : IDisposable
     {
         Task<Compatibility> GetCompatibilityAsync(string subject);
-        Task<Compatibility> GetGlobalCompatibilityAsync();
         Task<RegisteredSchema> GetLatestSchemaAsync(string subject);
         Task<Schema> GetSchemaAsync(int id, string format = null);
         Task<RegisteredSchema> GetSchemaAsync(string subject, int version);
         Task<List<string>> GetSubjectsAsync();
         Task<List<int>> GetSubjectVersionsAsync(string subject);
         Task<int> RegisterSchemaAsync(string subject, Schema schema, bool normalize);
-        Task<Config> SetCompatibilityAsync(string subject, Compatibility compatibility);
-        Task<Config> SetGlobalCompatibilityAsync(Compatibility compatibility);
+        Task<Compatibility> UpdateCompatibilityAsync(string subject, Compatibility compatibility);
         Task<bool> TestCompatibilityAsync(string subject, int versionId, Schema schema);
         Task<bool> TestLatestCompatibilityAsync(string subject, Schema schema);
         Task<RegisteredSchema> LookupSchemaAsync(string subject, Schema schema, bool ignoreDeletedSchemas, bool normalize);