+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# vNext
+
+## Enhancements
+
+- Added support for external JSON schemas in `JsonSerializer` and `JsonDeserializer` (#2042).
+- Added compatibility methods to CachedSchemaRegistryClient ([ISBronny](https://github.com/ISBronny), #2097).
+
+
 # 2.2.0
 
 ## Enhancements
@@ -21,7 +29,6 @@
 
 - References librdkafka.redist 2.1.1. Refer to the [librdkafka v2.1.1 release notes](https://github.com/confluentinc/librdkafka/releases/tag/v2.1.1) for more information.
 - Less heap allocations when calling Produce ([bjornbouetsmith](https://github.com/bjornbouetsmith), #2020)
-
 
 # 2.1.0